small changes v2
