+++ docker-compose.yml
@@ -1,37 +1,47 @@
-from sqlalchemy import create_engine, Column, Integer, String
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
-import datetime
-from dotenv import load_dotenv
-import os
-from sqlalchemy.dialects.postgresql import UUID
-import uuid
+version: '3'
 
-load_dotenv()
+services:
 
-DATABASE_URL = os.getenv("DATABASE_URL")
+  redis:
+    image: redis:alpine
+    ports:
+      - "6379:6379"
 
-Base = declarative_base()
+  report:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - "8000:8000"
+    depends_on:
+      - postgres
+    volumes:
+      - .:/usr/src/app
+    command: bash -c "wait-for-it 64.226.65.160:5432 -- uvicorn report_fastapi:app --host 0.0.0.0 --port 8000"
 
+  app:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    env_file:
+      - .env.docker
+    depends_on:
+      - postgres
+      - redis
+    volumes:
+      - .:/usr/src/app
+    command: bash -c "wait-for-it 64.226.65.160:5432 -- python app.py"
 
-class FinancialRecord(Base):
-    __tablename__ = 'financial_records'
+  postgres:
+    image: postgres:latest
+    environment:
+      POSTGRES_DB: financial_records
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+    ports:
+      - "5432:5432"
+    volumes:
+      - app_db:/var/lib/postgresql/data
 
-    message_id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, unique=True, nullable=False)
-    user_id = Column(Integer)
-    username = Column(String)
-    user_message = Column(String)
-    product = Column(String)
-    price = Column(Integer)
-    quantity = Column(Integer)
-    status = Column(String)
-    amount = Column(Integer)
-    timestamp = Column(String, default=lambda: (
-            datetime.datetime.utcnow() + datetime.timedelta(hours=6)
-    ).strftime('%d-%m-%y %H:%M'))
-
-
-engine = create_engine(DATABASE_URL, echo=True)
-Base.metadata.create_all(bind=engine)
-
-Session = sessionmaker(bind=engine)
+volumes:
+  app_db:
