+++ report_generator.py
@@ -1,67 +1,12 @@
-﻿aiofiles==23.1.0
-aiogram==2.15
-aiohttp==3.8.6
-aiosignal==1.3.1
-annotated-types==0.6.0
-anyio==3.7.1
-async-timeout==4.0.3
-attrs==23.1.0
-Babel==2.9.1
-certifi==2023.7.22
-charset-normalizer==3.3.0
-click==8.1.7
-colorama==0.4.6
-configure==0.5
-contourpy==1.2.0
-cycler==0.12.1
-dataclasses-json==0.6.1
-dill==0.3.7
-exceptiongroup==1.1.3
-fastapi==0.104.1
-fonttools==4.45.0
-frozenlist==1.4.0
-greenlet==3.0.0
-h11==0.14.0
-idna==3.4
-Jinja2==3.1.2
-jsonpatch==1.33
-jsonpointer==2.4
-kiwisolver==1.4.5
-langchain==0.0.312
-langsmith==0.0.43
-magic-filter==1.0.12
-MarkupSafe==2.1.3
-marshmallow==3.20.1
-matplotlib==3.8.2
-multidict==6.0.4
-mypy-extensions==1.0.0
-numexpr==2.8.7
-numpy==1.26.0
-openai==0.28.1
-packaging==23.2
-pandas==2.1.3
-Pillow==10.1.0
-psycopg2-binary==2.9.9
-pydantic==2.3.0
-pydantic_core==2.6.3
-pyparsing==3.1.1
-pyTelegramBotAPI==4.14.0
-python-dateutil==2.8.2
-python-dotenv==1.0.0
-pytz==2023.3.post1
-PyYAML==6.0.1
-redis==5.0.1
-reportlab==4.0.7
-requests==2.31.0
-six==1.16.0
-sniffio==1.3.0
-SQLAlchemy==2.0.21
-starlette==0.27.0
-tenacity==8.2.3
-tqdm==4.66.1
-typing-inspect==0.9.0
-typing_extensions==4.8.0
-tzdata==2023.3
-urllib3==2.0.6
-uvicorn==0.24.0.post1
-yarl==1.9.2
+﻿from jinja2 import Template
+
+
+def generate_html_report(financial_records):
+    with open("templates/report.html", "r") as template_file:
+        template_content = template_file.read()
+
+    template = Template(template_content)
+
+    html_content = template.render(financial_records=financial_records)
+
+    return html_content
